Run the migration with the next release
